requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic<2
pandas==2.2.3
numpy==1.26.4
sentence-transformers==2.7.0
scikit-learn==1.3.2
joblib==1.4.0
groq==0.31.0
pinecone-client==3.2.2
tqdm==4.66.5
requests>=2.32.3
